requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
Django==3.1.1
django-cloudinary-storage==0.3.0
django-utils-six==2.0
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
